+++ tests/Unit/Stubs/SMO.cs
@@ -484,6 +484,14 @@ namespace Microsoft.SqlServer.Management.Smo
             }
         }
 
+        public void Disable()
+        {
+        }
+
+        public string Certificate;
+        public string AsymmetricKey;
+        public string Language;
+
         public void Drop()
         {
         }